Write chapter on doubles
This covers some of the reasons why you would want to test in isolation
and how to use test doubles to achieve that goal.

This section should eventually cover mocking, stubbing, and spies as
other tools for testing in isolation. The introduction, "Benefits", and
"Pragmatic Approach" sections may need to be reworked a bit as those
topics get added.