Pull in dependencies for release script
We cannot assume the following dependencies of the release script to be
installed already:

* `../vbump/bin/vbump`
* GNU `sort`
* GNU `sed`

`vbump` is a utility for intelligently bumping version numbers. The
release script assumed that  the `vbump` project would be cloned onto
the author's machine as a sibling of the book repo.

Instead, I copied the script as `bin/vbump` to distribute it with the
other utilities of the book repo. Unfortunately, the script required
versions of `sort` and `sed` that were not installed by default on OSX.

They can be downloaded on OSX via the homebrew `coreutils` package as
`gsort` and `gsed`. I modified the script to use these and to attempt to
install `coreutils` via homebrew if they are not detected on the system.

This will likely break on Linux machines unless they have a `gsort`
command that points to the correct version of GNU `sort`