Add chapter on stubbing
This adds a section on how to use RSpec's stubbing methods (`allow`,
`to_receive`, and `and_return`) to isolate collaborators that are not
passed in.

Some of the downsides of stubbing (such as brittleness) will be explored
in a later chapter on the downsides and gotchas of testing in isolation
in general.

https://trello.com/c/ddSjoQ7i