Rename `FactoryGirl` to `FactoryBot`
The `FactoryGirl` gem was
[renamed](https://robots.thoughtbot.com/factory_bot) to `FactoryBot`.
This commit goes through and replaces all references to `FactoryGirl`
and `factory_girl` in the book and example app.

Some parts of the book refer to older commits in the example app. In
these places, I've added a note saying the gem was formerly called
`FactoryGirl`.